Tidy up discussion about event names (#321)
* Tidy up discussion about event names

This section had some hangover text from when an event contained
separate category and type fields. That's confusing.

Lets be clearer and more explicity about how we expect the name field to
be used. We should probably add additional language about what to do if
a parser encouners a name field *is not* valid. But that can be done as
followup.

Co-authored-by: Robin Marx <rmarx@akamai.com>