Reflow of main schema document first half
The organic growth of the document has made, in my opinion, the flow
of the document a bit janky.

This is especially so for JSON-SEQ. After reading the whole document
that frames things as qlog - there's a text buried down in the
serialization that requires the reader to discard a lot of internalized
context. It shouldn't be like that because QLogFile and QLogFileSeq are
separate things that share a tonne of qlog types.

This change focuses on sections 3-5. We draw our the primary themes that
underpin qlog by picking out the terms that we actually use. Instead of
schema -> traces -> Field name semantics in one section, we move to
QLogFile schema and related traces and Tace, QlogFileSeq and related TraceSeq,
VantagePoint, Events, Raw packet info, and Common Events. Finishing up with
guidelines for people to define their own events.