proposal: remove QPACK events
We've heard some feedback that the QPACK events are too low-level to be useful
and that higher-level QPACK events would be preferred. However, in the absence of
a proposal for higher-level QPACK events we are stuck with what is currently
defined ... and nobody seems to be using these.

Proposal: remove low-level QPACK events from the spec so that consensus and
implementation experience of the remaining HTTP/3 events can be completed.