Reframe JSON interop concerns seperately of I-JSON (#360)
Since we added JSON-SEQ, it's been hard to square the circle about
where I-JSON fits in the document structure. I think that's because
actually the discussion should be framed in the context of interop
issues related to JSON messages. Our 64-bit values pose a problem
and the way CDDL suggest to solve that doesn't work - so we lay it
all out here.

There is obvious overlap here with PR #359. Whichever lands first
(if they even do) will cause some conflict that is easy to fixup.

Closes #339