Add fin and reset_stream fields
Fixes #375.

Per the discussion, it can be useful to know when an application triggers a
stream close, or when it receives one, without having to do offset/length maths.

I omitted stop_sending in this commit because it seems a little bit of a stretch,
mainly because it's wire image has no offset or final size. Also, STOP_SENDING
relates to asking the peer to do something, which further complicates the
(potential) use of offset and length fields.

An alternative would be to keep `fin` here but define new events for reset and
stop. Those could both contain from, to and error fields but only the reset
would have offset and length.