Tweak stream_state_updated
This implements my suggestions from #374, adding a new optional
initiator field to supplement the stream_type field, along with
making the stream_side field mandatory.

Having just implemented a prototype pcap to qlog tool, that tool
needs to try and build an understanding of who is a QUIC server
and client, and in what direction messages are flowing. So even
an observer vantage point can determine and log `stream_side`.

Closes #374